# api-gateway/swagger.yaml
openapi: 3.0.0
info:
  title: ALT_LAS API Gateway
  description: API Gateway for ALT_LAS project - Platform designed to manage computer systems with artificial intelligence. Handles authentication, routing, rate limiting, and service discovery.
  version: 1.1.0 # Version updated
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.alt-las.com
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Access Token required for protected endpoints.
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
          description: Error message
        error:
          type: string
          description: Detailed error information (optional, development only)

    # Auth Schemas
...
    ArchiveRequest:
      type: object
      required:
        - lastFile
      properties:
        lastFile:
          type: string
          description: Path or identifier of the *.last file to archive
        metadata:
          type: object
          additionalProperties: true
          description: Additional metadata for archiving
    ArchiveResponse:
      type: object
      properties:
        id:
          type: string
          description: Unique ID for the archive entry
        status:
          type: string
          enum: [success, error]
          description: Status of the archiving process
        atlasId:
          type: string
          description: Identifier within the *.atlas archive system
        successRate:
          type: number
          minimum: 0
          maximum: 100
          description: Calculated success rate based on archived data
        metadata:
          type: object
          properties:
            userId:
              type: string
              format: uuid
              description: ID of the user related to the archived task
paths:
...
  /api/v1/archive:
    post:
      summary: Result archiving
      description: Archives the *.last file and saves it to the *.atlas database. Proxies request to the Archive Service.
      tags: [Microservices V1]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArchiveRequest'
      responses:
        '200':
          description: Successful archiving (mock response)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArchiveResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized (Missing or invalid token)
        '500':
          description: Internal server error or Archive Service error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Archive Service unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/archive/{id}:
    get:
      summary: Query archive record
      description: Queries a specific archive record.
      tags: [Microservices V1]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: The unique ID of the archive record.
      responses:
        '200':
          description: Archive record (mock response)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArchiveResponse'
        '401':
          description: Unauthorized (Missing or invalid token)
        '404':
          description: Archive record not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...